+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Improve metrics label/tooltip formatting
+- Fix metrics chart not rendering
 - Updated Predis to v1.1.10
 - Support `composer/installers` v1 and v2