+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Improve metrics label/tooltip formatting
 - Fix metrics chart not rendering
 - Updated Predis to v1.1.10
+- Updated Credis to v1.13.0
 - Support `composer/installers` v1 and v2
 - Link to settings page when foreign drop-in was found
 - Added `wp_cache_flush_runtime()` function